feat(session): refactor session handling to use SessionService and update repository methods for consistency

+++ GO-Manager/internal/database/postgres/session_repository/delete.go
@@ -1,7 +1,6 @@
 package session_repository
 
-
-func (r *Repository) Delete(sessionId int) error {
+func (r *Repository) Delete(sessionId int64) error {
 	_, err := r.db.Exec(`
 		DELETE FROM sessions
 		WHERE id = $1
